requirements.txt
streamlit
plotly
geopandas
shapely
polyline
concurrent.futures
numpy